+++ test/e2e/gameserver_test.go
@@ -175,14 +175,10 @@ func TestUnhealthyGameServersWithoutFreePorts(t *testing.T) {
 	}
 
 	newGs, err := gameServers.Create(gs.DeepCopy())
-	assert.Nil(t, err)
+	assert.NoError(t, err)
 
-	_, err = framework.WaitForGameServerState(newGs, v1alpha1.GameServerStateUnhealthy, 10*time.Second)
-	assert.NotNil(t, err)
-
-	_, err = gameServers.Get(newGs.Name, metav1.GetOptions{})
-	assert.NotNil(t, err)
-	assert.True(t, k8serrors.IsNotFound(err))
+	_, err = framework.WaitForGameServerState(newGs, v1alpha1.GameServerStateUnhealthy, time.Minute)
+	assert.NoError(t, err)
 }
 
 func TestGameServerUnhealthyAfterDeletingPod(t *testing.T) {
@@ -208,27 +204,7 @@ func TestGameServerUnhealthyAfterDeletingPod(t *testing.T) {
 	err = podClient.Delete(pod.ObjectMeta.Name, nil)
 	assert.NoError(t, err)
 
-	// TODO [markmandel@google.com]: Should GameServers that are Unhealthy and not in a fleet be deleted?
-	err = wait.PollImmediate(2*time.Second, time.Minute, func() (bool, error) {
-		gs, err := framework.AgonesClient.StableV1alpha1().GameServers(readyGs.Namespace).Get(readyGs.ObjectMeta.Name, metav1.GetOptions{})
-
-		// just in case
-		if k8serrors.IsNotFound(err) {
-			return true, nil
-		}
-
-		if err != nil {
-			logrus.WithError(err).Warn("error retrieving gameserver")
-			return false, nil
-		}
-
-		if gs.Status.State == v1alpha1.GameServerStateUnhealthy {
-			return true, nil
-		}
-
-		return false, nil
-	})
-
+	_, err = framework.WaitForGameServerState(readyGs, v1alpha1.GameServerStateUnhealthy, time.Minute)
 	assert.NoError(t, err)
 }
 
@@ -309,6 +285,35 @@ func TestGameServerSelfAllocate(t *testing.T) {
 		}
 
 		return gs.Status.State == v1alpha1.GameServerStateAllocated, nil
+	})
+
+	assert.NoError(t, err)
+}
+
+func TestGameServerShutdown(t *testing.T) {
+	t.Parallel()
+	gs := defaultGameServer()
+	readyGs, err := framework.CreateGameServerAndWaitUntilReady(defaultNs, gs)
+	if err != nil {
+		t.Fatalf("Could not get a GameServer ready: %v", err)
+	}
+	assert.Equal(t, readyGs.Status.State, v1alpha1.GameServerStateReady)
+
+	reply, err := e2eframework.SendGameServerUDP(readyGs, "EXIT")
+	if err != nil {
+		t.Fatalf("Could not message GameServer: %v", err)
+	}
+
+	assert.Equal(t, "ACK: EXIT\n", reply)
+
+	err = wait.PollImmediate(time.Second, time.Minute, func() (bool, error) {
+		gs, err = framework.AgonesClient.StableV1alpha1().GameServers(defaultNs).Get(readyGs.ObjectMeta.Name, metav1.GetOptions{})
+
+		if k8serrors.IsNotFound(err) {
+			return true, nil
+		}
+
+		return false, err
 	})
 
 	assert.NoError(t, err)